module-lattice: rename `encoding` and `utils`
Renames the following:
- `module_lattice::encode` => `module_lattice::encoding`
- `module_lattice::util` => `module_lattice::utils`

These are more conventional names within the scope of the @RustCrypto
project.
