+++ module-lattice/src/lib.rs
@@ -14,8 +14,8 @@
 pub mod algebra;
 
 /// Packing of polynomials into coefficients with a specified number of bits.
-pub mod encode;
+pub mod encoding;
 
 /// Utility functions such as truncating integers, flattening arrays of arrays, and unflattening
 /// arrays into arrays of arrays.
-pub mod util;
+pub mod utils;
